requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
black==25.9.0
certifi==2025.10.5
cfgv==3.4.0
click==8.3.0
Deprecated==1.2.18
distlib==0.4.0
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
fastapi==0.118.2
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
filelock==3.20.0
flake8==7.3.0
frozenlist==1.8.0
h11==0.14.0
httpcore==0.16.3
httptools==0.6.4
httpx==0.23.3
identify==2.6.15
idna==3.10
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
multidict==6.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
orjson==3.11.3
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.3.0
propcache==0.4.1
pycodestyle==2.14.0
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
PyYAML==6.0.3
rfc3986==1.5.0
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
sentry-sdk==2.40.0
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
starlette==0.48.0
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
virtualenv==20.35.3
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
